refactoring to make underlying server replaceable
